+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Export unsigned PSBT and import signed PSBT in watch-only wallets
+
 ### Changed
 - Hide sensitive info in screenshots